return `expected fail' if runs on non-filesystem node, according to the report [gfarm-developers:02489] by Yamamoto-san
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@3268 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ regress/gfs_hook/tool/dd/dd.sh
@@ -6,6 +6,13 @@ trap 'rm -f $hooktmp; exit $exit_trap' $trap_sigs
 
 ulimit -c 0	# do not dump core
 
+if ! gfhost -M `hostname` >/dev/null
+then
+	# dd issues `Socket operation on non-socket' on non-filesystem node,
+	# this problem is not documented yet
+    	exit $exit_xfail
+fi    
+
 if dd if=$data/1byte of=$hooktmp
 then
 	exit_code=$exit_pass
